doc: Adding more references to EdgeVPN network token format in getting started and Installation section
At my first view of the getting started section, I was confused by what the QR Code was. And I was confused by the term TOKEN or NETWORK_TOKEN which was use but not necessarly explained in the getting started or Installation section. Therefore I propose to add a few explanation on the EdgeVPN token in those sections which are the first sections that a reader reads.

Signed-off-by: Nicolas-Peiffer <102670102+Nicolas-Peiffer@users.noreply.github.com>