Quickstart: extend, lifecycle and TB (#507)
* add hadron quickstart for trusted boot

* readability

* objective and description

* extending system and trusted boot changes

* improve wording

* Add alert for Hadron toolchain in Dockerfile quickstart

This update includes an alert explaining the use of the Hadron `hadron-toolchain` image as a build stage, highlighting its role in providing a complete build environment while maintaining security and size considerations for production images.

* Update button styles in quickstart documentation

Changed button classes from 'btn-outline-primary' to 'btn-primary' for improved visibility and consistency in the quickstart section, specifically in the links for extending the system and trusted boot.

* move upgrade into its own short quickstart

* add rest of wording for lifecycle management quickstart

* updates to lifecycle management

* Refine Kubernetes installation instructions in Dockerfile quickstart

Updated the installation commands to remove the model specification and clarified the ability to choose specific Kubernetes versions using the `--provider` flag. Adjusted the FAQ section to reflect these changes, enhancing clarity for users.

* remove section

Signed-off-by: Mauro Morales <contact@mauromorales.com>

* Improve quickstart

* consistent

* beta2

* seo updates

* typo

* small issues found while running the scripts

* Update trusted-boot.md to use auroraboot version 0.16.2 in example commands

* Update video links in quickstart documentation for consistency and clarity

* move link to systemd extensions to the top

---------

Signed-off-by: Mauro Morales <contact@mauromorales.com>